requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
django-environ==0.7.0
django-cors-headers
djangorestframework-simplejwt
psycopg2>=2.8.6,<2.9